#2 unit test
